+++ exo-gadget-pack/gadget-pack-services/src/main/java/org/exoplatform/platform/gadget/services/LoginHistory/LoginHistoryService.java
@@ -40,7 +40,7 @@ public interface LoginHistoryService {
 	
 	public Set<String> getLastUsersLogin(long fromTime) throws Exception;
 	
-	public boolean isActiveUser(String userId, int days);
+	public boolean isActiveUser(String userId, int days) throws Exception;
 	
 	public Map<String, Integer> getActiveUsers(long fromTime);
